+++ src/host/emscripten/mod.rs
@@ -15,8 +15,8 @@ use crate::{
     BufferSize, BuildStreamError, Data, DefaultStreamConfigError, DeviceDescription,
     DeviceDescriptionBuilder, DeviceId, DeviceIdError, DeviceNameError, DevicesError,
     InputCallbackInfo, OutputCallbackInfo, PauseStreamError, PlayStreamError, SampleFormat,
-    SampleRate, StreamConfig, StreamError, SupportedBufferSize, SupportedStreamConfig,
-    SupportedStreamConfigRange, SupportedStreamConfigsError,
+    SampleRate, StreamConfig, StreamError, StreamInstant, SupportedBufferSize,
+    SupportedStreamConfig, SupportedStreamConfigRange, SupportedStreamConfigsError,
 };
 
 // The emscripten backend currently works by instantiating an `AudioContext` object per `Stream`.
@@ -278,6 +278,10 @@ impl StreamTrait for Stream {
         });
         Ok(())
     }
+
+    fn now(&self) -> StreamInstant {
+        StreamInstant::from_secs_f64(self.audio_ctxt.current_time())
+    }
 }
 
 fn audio_callback_fn<D>(
@@ -299,15 +303,13 @@ where
             let data = temporary_buffer.as_mut_ptr() as *mut ();
             let mut data = unsafe { Data::from_parts(data, len, sample_format) };
             let now_secs: f64 = audio_ctxt.current_time();
-            let callback = crate::StreamInstant::from_secs_f64(now_secs);
+            let callback = StreamInstant::from_secs_f64(now_secs);
             // TODO: Use proper latency instead. Currently, unsupported on most browsers though, so
             // we estimate based on buffer size instead. Probably should use this, but it's only
             // supported by firefox (2020-04-28).
             // let latency_secs: f64 = audio_ctxt.outputLatency.try_into().unwrap();
             let buffer_duration = frames_to_duration(len, sample_rate as usize);
-            let playback = callback
-                .add(buffer_duration)
-                .expect("`playback` occurs beyond representation supported by `StreamInstant`");
+            let playback = callback + buffer_duration;
             let timestamp = crate::OutputStreamTimestamp { callback, playback };
             let info = OutputCallbackInfo { timestamp };
             data_callback(&mut data, &info);